修改 1.作廢提示 2.起日止日 3.車主編輯

+++ package/src/app/(DashboardLayout)/vehicle-management/components/table/PolicyManagementTable.js
@@ -29,6 +29,11 @@ const PolicyManagmentTable = ({ data, carLicenseNum, refreshData }) => {
   };
 
   const handleDeleteClick = (carLicenseNum, insuranceCardNum) => {
+    const isConfirmed = window.confirm(
+      "確定要作廢這筆資料嗎？此操作無法還原！"
+    );
+    if (!isConfirmed) return;
+
     deleteInsurance(
       { carLicenseNum, insuranceCardNum },
       {
